Remove casting to int of some vars, causes issues with 32 bit versions of php.

+++ nzedb/Binaries.php
@@ -367,10 +367,10 @@ class Binaries
 				}
 			}
 
-			$left = (int)$this->messagebuffer;
-			$last = (int)($grouplast = $data['last'] - $left);
+			$left = $this->messagebuffer;
+			$last = ($grouplast = $data['last'] - $left);
 		} else {
-			$first = (int)$groupArr['last_record'];
+			$first = $groupArr['last_record'];
 
 			// Leave 50%+ of the new articles on the server for next run (allow server enough time to actually make parts available).
 			$newcount = $data['last'] - $first;
@@ -378,7 +378,7 @@ class Binaries
 			if ($newcount > $this->messagebuffer) {
 				// Drop the remaining plus $this->messagebuffer, pick them up on next run
 				if ($newcount < (2 * $this->messagebuffer)) {
-					$left = ((int) ($newcount / 2));
+					$left = (($newcount / 2));
 					$last = $grouplast = ($data['last'] - $left);
 				} else {
 					$remainingcount = $newcount % $this->messagebuffer;
@@ -386,7 +386,7 @@ class Binaries
 					$last = $grouplast = ($data['last'] - $left);
 				}
 			} else {
-				$left = ((int) ($newcount / 2));
+				$left = (($newcount / 2));
 				$last = $grouplast = ($data['last'] - $left);
 			}
 		}
